requirements.txt
streamlit==1.36.0             # For building the UI
chromadb==0.5.3               # Vector database to store and search embeddings
pydantic<2                    # Needed for Chroma compatibility
sentence-transformers==2.7.0  # Pre-trained embedding models
langchain==0.2.10             # To connect retriever + LLM
langchain-community==0.2.9    # Community integrations (Chroma, etc.)
pypdf==4.3.1                  # Extract text from PDFs
markdown-it-py==3.0.0         # Rendering markdown if needed
rapidfuzz==3.9.6              # For fuzzy matching (optional term expansion)
numpy                         # Math operations for embeddings
tqdm                          # Progress bars during ingestion
openai
python-dotenv
